ParaCrawl: switch from Bicleaner to Zipporah filtering
ParaCrawl has two heuristics for determining if something is a
reasonable translation: Bicleaner and Zipporah. Each of them outputs a
rating for the text, and which file you download affects whether the
threshold for inclusion is the Bicleaner rating or the Zipporah rating.

The Bicleaner version was giving me a bunch of bad data, consisting of
something like gzip files interpreted as Windows-1252. All of these
entries are marked with very low scores by Zipporah. I believe I should
switch to Zipporah.

The ParaCrawl developers (whoever they are) seem more fond of Zipporah
at the moment, as well.